+++ game/chess.py
@@ -6,9 +6,17 @@ class Chess:
         self.__board__ = Board()
         self.__turn__ = "WHITE"
 
-    # def move(self,from_row, from_col, to_row, to_col):
-    #     #piece = self.__board__.get_piece(from_row, from_col)
-    #     self.change_turn()
+    def move(self,from_row, from_col, to_row, to_col):
+        piece = self.__board__.get_piece(from_row, from_col)
+
+        if piece == "No piece":
+            print("No puede mover pieza que no esta")
+            return
+
+
+        self.__board__.move_piece(from_row, from_col, to_row, to_col)
+        self.change_turn()
+
 
     def change_turn(self):
         if self.__turn__ == "WHITE":