+++ test/test.py
@@ -1,12 +1,8 @@
 import unittest
 
-from unittest.mock import patch
-from io import StringIO
-
-from game.rook import Rook, Pawn
+from game.piece import Rook, Pawn
 from game.board import Board
 from game.chess import Chess
-from game.main import main
 
 class TestChess(unittest.TestCase):
     def setUp(self):
@@ -24,19 +20,53 @@ class TestChess(unittest.TestCase):
         self.chess.change_turn()
         self.assertEqual(self.chess.__turn__, "WHITE")
 
+
+    def test_move_piece(self):
+        self.chess.move(7, 0, 6, 0)
+        
+        self.assertEqual(self.chess.__board__.get_piece(7, 0), "No piece")
+
+        self.assertEqual(self.chess.__turn__, "BLACK")
+
+    def test_move_no_piece(self):
+
+
+        self.assertEqual(self.chess.__board__.get_piece(4, 4), "No piece")
+        self.assertEqual(self.chess.__board__.get_piece(3, 3), "No piece")
+
+        self.assertEqual(self.chess.__turn__, "WHITE")
+
+
 class TestBoard(unittest.TestCase):
-    def test_board_init(self):
-        # Test that Rooks are correctly placed
-        board = Board()
-        self.assertEqual(board.get_piece(0, 0), "ROOK (BLACK)")
-        self.assertEqual(board.get_piece(7, 0), "ROOK (WHITE)")
+    def setUp(self):
+        self.board = Board()
+
+    def test_init_board(self):
+        self.assertEqual(self.board.__positions__[0][0].__type__, "ROOK")
         
-        # Test that Pawns are correctly placed
-        self.assertEqual(board.get_piece(1, 0), "PAWN (BLACK)")
-        self.assertEqual(board.get_piece(6, 1), "PAWN (WHITE)")
+
+    def test_get_piece_empty(self):
+        self.assertEqual(self.board.get_piece(3, 3), "No piece")
+
+    def test_move_piece(self):
+        self.board.move_piece(0, 0, 0, 1)
         
-        # Test an empty square
-        self.assertEqual(board.get_piece(4, 4), "No piece")
+
+        self.assertEqual(self.board.get_piece(0, 1), ('La pieza de esa posicion es: ', {'ROOK'}, {'BLACK'}))
+
+        self.assertEqual(self.board.get_piece(0, 0), "No piece")
+
+    def test_move_piece_no_piece(self):
+
+        result = self.board.move_piece(3, 3, 4, 4)
+
+
+        self.assertEqual(result, "No piece to move")
+
+        self.assertEqual(self.board.get_piece(3, 3), "No piece")
+        self.assertEqual(self.board.get_piece(4, 4), "No piece")
+
+
 
 class TestRook(unittest.TestCase):
     def test_rook_init(self):
@@ -50,24 +80,24 @@ class TestRook(unittest.TestCase):
         self.assertEqual(pawn.__type__, "PAWN")
 
 
-class TestMain(unittest.TestCase):
-    @patch('builtins.input', side_effect=[0, 0, 1, 0])
-    @patch('sys.stdout', new_callable=StringIO)  # Captura la salida estándar (print)
-    def test_main_valid_move(self, mock_stdout, mock_input):
-        main()  # Ejecuta la función main()
+# class TestMain(unittest.TestCase):
+    # @patch('builtins.input', side_effect=[0, 0, 1, 0])
+    # @patch('sys.stdout', new_callable=StringIO)  # Captura la salida estándar (print)
+    # def test_main_valid_move(self, mock_stdout, mock_input):
+    #     main()  # Ejecuta la función main()
 
-        # Verifica la salida capturada
-        output = mock_stdout.getvalue()
-        self.assertIn("The piece you have choosen is: ROOK (BLACK)", output)
+    #     # Verifica la salida capturada
+    #     output = mock_stdout.getvalue()
+    #     self.assertIn("The piece you have choosen is: ROOK (BLACK)", output)
 
-    @patch('builtins.input', side_effect=[3, 3, 4, 4])
-    @patch('sys.stdout', new_callable=StringIO)  
-    def test_main_no_piece(self, mock_stdout, mock_input):
-        main()  # Ejecuta la función main()
+    # @patch('builtins.input', side_effect=[3, 3, 4, 4])
+    # @patch('sys.stdout', new_callable=StringIO)  
+    # def test_main_no_piece(self, mock_stdout, mock_input):
+    #     main()  # Ejecuta la función main()
 
-        # Verifica la salida capturada
-        output = mock_stdout.getvalue()
-        self.assertIn("The piece you have choosen is: No piece", output)
+    #     # Verifica la salida capturada
+    #     output = mock_stdout.getvalue()
+    #     self.assertIn("The piece you have choosen is: No piece", output)