+++ changelog.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+#--------------Primer Sprint----------------
+## [0.12] - 15-8-24
+Added funcion move to chess, move_piece method to board, and fixed get_piece method
+
 ## [0.11] - 14-8-24
 Added get_piece method to board, and the pawns on the board